Add Docker client components: Implement ImageClient, SystemClient, and VolumeClient for managing Docker resources. Introduce SearchInput widget for enhanced user interaction in the Testronaut DB Browser.
